Robin: Reworking Dalvik Heap Memory configuration
This patch removes all occurences of touching
the Dalvik Heap Memory configuration.

This is done to implement the optimal
Memory configuration for this device which is
documented below.

This is the previous configuration which is the
generic one by Qualcomm:
heapstartsize: 8m
heapgrowthlimit: 192m
heapsize: 512m
heaptargetutilization: 0.75
heapminfree: 512k
heapmaxfree: 8m

The values relevant for the proportional
dalvik memory heap check are heapgrowthlimit,
heapsize and heaptargetutilization.

The previous configuration passes the
proportional dalvik memory heap check.

In the following we see the new dalvik
heap memory values which were evaluated
dynamically for the Nextbit Robin while
considering that it is a device with 3GB
RAM:

heapstartsize: 8m
heapgrowthlimit: 288m
heapsize: 768m
heaptargetutilization: 0.75
heapminfree: 1m
heapmaxfree: 8m

The values that were modified are heapgrowthlimit,
heapsize and heapminfree.

Below we see the check:

heaptargetutilization/2:
0.75/2 = 0.375

To pass the check this has to be true:
0.375 * heapsize = heapgrowthlimit

-> 0.375 * 768 = 288

That statement is true, that means we
are still passing the proportional
dalvik memory heap check.

Additionally to the values which were modified
to be optimal for Nextbit Robin, heapminfree
was changed from 512kb to 1m.
This is to fix the inconsistent and poor
GC performance which was noticed.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>