[PARANOID-62] common: power-8992: Increase timer_rate when screen is off
This patch highers the interactive timer_rate
for little cores to 50ms while the screen is off.

The result is that cpu load is recalculated less
often which results in more conservative cpu scaling.

This is unneeded to do on big cores, since
big cores are offline when the screen is off.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>