[PARANOID-63] common: Reduce read-ahead value to 128kb
Qualcomm decided to increase the read-ahead
value to 512kb [1].

This value resulted in frequent cases
where read-ahead would be waiting for new
I/O requests and thus ending up being synchronous
rather than asynchronous.

As result, the main goal of Qualcomm's patch
[1] which is to increase throughput ends
up in a regression where read throughput
is significantly lower than it ever
would be with the recommended [2]
read-ahead value.

Overall we can say that a value of
512kb is far from optimal on a device
with EMMC.

This patch uses our existing loop
in ramdisk to set the read-ahead value
to 128kb.

[1] https://source.codeaurora.org/quic/la/kernel/msm-3.10/commit/include/linux/mm.h?h=LA.BF64.1.2.3-01310-8x92.0&id=403678af49ff8f8ad050407f62e5f8918b1825a4
[2] https://github.com/torvalds/linux/blob/v4.9/include/linux/mm.h#L2109

Suggested-by: Jake Weinstein <xboxlover360@gmail.com>
Signed-off-by: Alex Naidis <alex.naidis@linux.com>